[3787] Move methods to the utils module
