--- stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/utils/extensions/Date.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2014-2022 Stream.io Inc. All rights reserved.
- *
- * Licensed under the Stream License;
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.getstream.sdk.chat.utils.extensions
-
-import io.getstream.chat.android.core.internal.InternalStreamChatApi
-import java.util.Date
-
-private const val ONE_MINUTE_IN_MILISECONDS = 60000
-
-@InternalStreamChatApi
-public fun Date.isInLastMinute(): Boolean = (Date().time - ONE_MINUTE_IN_MILISECONDS < time)
