+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/extensions/User.kt
@@ -18,10 +18,10 @@ package io.getstream.chat.android.ui.common.extensions
 
 import android.content.Context
 import android.text.format.DateUtils
-import com.getstream.sdk.chat.utils.extensions.isInLastMinute
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.common.extensions.internal.EMPTY
+import io.getstream.chat.android.uiutils.extensions.isInLastMinute
 
 public fun User.getLastSeenText(context: Context): String {
     if (online) {
