+++ stream-chat-android-ui-utils/build.gradle
@@ -52,4 +52,5 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 }
 
 dependencies {
+    implementation project(':stream-chat-android-client')
 }
